+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.0 Under development
 
 - Enh #8: Refactor codebase to improve performance (@terabytesoftw)
+- Bug #9: Enhance documentation and improve clarity across HTML helper classes (@terabytesoftw)
 
 ## 0.2.0 March 30, 2024