+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enh #8: Refactor codebase to improve performance (@terabytesoftw)
 - Bug #9: Enhance documentation and improve clarity across HTML helper classes (@terabytesoftw)
+- Bug #10: Improve pattern handling and enhance test coverage for `Naming` class (@terabytesoftw)
 
 ## 0.2.0 March 30, 2024